+++ batcomputer.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-"""Call example's CLI."""
-
-from example.batcomputer_cli import CLI
-
-CLI()