+++ cly/utils.py
@@ -3,11 +3,31 @@
 # Scripts that manipulate the shell must always be careful with possible
 # security implications.
 import subprocess  # nosec
-from typing import List, Optional, Union
+from pathlib import Path
+from typing import List, Optional, Sequence, Tuple, Union
 
 from .colors import color_text
 
 SPACE = " "
+
+
+def print_error_message(command: subprocess.Popen) -> None:  # type: ignore
+    """
+    Print error message from a command.
+
+    Parameters
+    ----------
+    command : subprocess.Popen
+        Command to get information from.
+
+    """
+    print(
+        color_text(
+            f"ERROR: Command '{command.args}' returned "  # type: ignore
+            f"non-zero exit status {command.returncode}.",
+            "red",
+        )
+    )
 
 
 def parse_arguments(arguments: Union[str, List[str]]) -> str:
@@ -31,8 +51,8 @@ def parse_arguments(arguments: Union[str, List[str]]) -> str:
 
 
 def get_output(
-    arguments: Union[str, List[str]]
-) -> subprocess.CompletedProcess:
+    arguments: Union[str, List[str]], directory: Optional[Path] = None
+) -> subprocess.CompletedProcess:  # type: ignore
     """
     Get the output information of the shell command.
 
@@ -42,6 +62,8 @@ def get_output(
     ----------
     arguments : Union[str, List[str]]
         A string, or list of strings, containing the commands and arguments.
+    directory : Optional[pathlib.Path]
+        Directory to run shell command, by default runs in current directory.
 
     Returns
     -------
@@ -49,17 +71,19 @@ def get_output(
         Command's output information.
 
     """
-    command = parse_arguments(arguments)
     return subprocess.run(
-        command,
+        parse_arguments(arguments),
         shell=True,  # nosec
         check=False,
         capture_output=True,
         encoding="utf-8",
+        cwd=directory,
     )
 
 
-def get_returncode(arguments: Union[str, List[str]]) -> int:
+def get_returncode(
+    arguments: Union[str, List[str]], directory: Optional[Path] = None
+) -> int:
     """
     Get the returncode of the shell command.
 
@@ -69,6 +93,8 @@ def get_returncode(arguments: Union[str, List[str]]) -> int:
     ----------
     arguments : Union[str, List[str]]
         A string, or list of strings, containing the commands and arguments.
+    directory : Optional[pathlib.Path]
+        Directory to run shell command, by default runs in current directory.
 
     Returns
     -------
@@ -76,12 +102,13 @@ def get_returncode(arguments: Union[str, List[str]]) -> int:
         Command's returncode.
 
     """
-    output = get_output(arguments)
-    return output.returncode
+    return get_output(arguments, directory).returncode
 
 
 def get_standard_output(
-    arguments: Union[str, List[str]], lines: bool = False
+    arguments: Union[str, List[str]],
+    lines: bool = False,
+    directory: Optional[Path] = None,
 ) -> Optional[List[str]]:
     """
     Get the standard output of the shell command.
@@ -95,6 +122,8 @@ def get_standard_output(
     lines : bool
         Separate output in lines instead of separating in words, by default
         False.
+    directory : Optional[pathlib.Path]
+        Directory to run shell command, by default runs in current directory.
 
     Returns
     -------
@@ -102,7 +131,7 @@ def get_standard_output(
         A list of strings containing the output's words or lines; else, None.
 
     """
-    output = get_output(arguments).stdout
+    output = get_output(arguments, directory).stdout
     if output:
         if lines:
             return [line for line in output.split("\n") if line]
@@ -112,9 +141,45 @@ def get_standard_output(
     return None
 
 
+def get_command_after_execution(
+    arguments: Union[str, List[str]], directory: Optional[Path] = None
+) -> subprocess.Popen:  # type: ignore
+    """
+    Get shell command information after it executes.
+
+    The command output is printed to the user.
+
+    Parameters
+    ----------
+    arguments : Union[str, List[str]]
+        Shell command and arguments to be executed.
+    directory : Optional[pathlib.Path]
+        Directory to run shell command, by default runs in current directory.
+
+    Returns
+    -------
+    subprocess.Popen
+        Executed command.
+
+    """
+    command = subprocess.Popen(
+        parse_arguments(arguments),
+        shell=True,  # nosec
+        encoding="utf-8",
+        cwd=directory,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+    )
+    with command:
+        while command.poll() is None:
+            print(command.stdout.readline(), end="")  # type: ignore
+        print(command.stdout.read(), end="")  # type: ignore
+    return command
+
+
 def run_command(
-    arguments: Union[str, List[str]]
-) -> subprocess.CompletedProcess:
+    arguments: Union[str, List[str]], directory: Optional[Path] = None
+) -> None:
     """
     Run the shell command.
 
@@ -124,11 +189,8 @@ def run_command(
     ----------
     arguments : Union[str, List[str]]
         A string, or list of strings, containing the commands and arguments.
-
-    Returns
-    -------
-    subprocess.CompletedProcess
-        Executes the command.
+    directory : Optional[pathlib.Path]
+        Directory to run shell command, by default runs in current directory.
 
     Raises
     ------
@@ -136,11 +198,40 @@ def run_command(
         If command fails.
 
     """
-    command = parse_arguments(arguments)
-    try:
-        return subprocess.run(
-            command, shell=True, check=True, encoding="utf-8"  # nosec
-        )
-    except subprocess.CalledProcessError as error:
-        print(color_text(f"ERROR: {error}", "red"))
-        raise SystemExit(error.returncode) from error
+    command = get_command_after_execution(arguments, directory)
+    if command.returncode:
+        print_error_message(command)
+        raise SystemExit(command.returncode)
+
+
+def run_multiple_commands(
+    commands: Sequence[Tuple[Union[str, List[str]], Optional[Path]]]
+) -> None:
+    """
+    Run multiple shell commands.
+
+    **Be careful about security implications when manipulating the shell!**
+
+    Parameters
+    ----------
+    commands: Sequence[Tuple[Union[str, List[str]], Optional[pathlib.Path]]]
+        List of commands, where each command is a tuple of commands and
+        arguments and directory, to be executed.
+
+    Raises
+    ------
+    SystemExit
+        If one of the command fails.
+
+    """
+    executed_commands = [
+        get_command_after_execution(arguments, directory)
+        for arguments, directory in commands
+    ]
+    error_commands = [
+        print_error_message(command)  # type: ignore
+        for command in executed_commands
+        if command.returncode
+    ]
+    if error_commands:
+        raise SystemExit(len(error_commands))