+++ docs/conf.py
@@ -32,7 +32,9 @@ nitpick_ignore = [
     ("py:class", "argparse.Namespace"),
     ("py:class", "argparse._ArgumentGroup"),
     ("py:class", "argparse._SubParsersAction"),
+    ("py:class", "pathlib.Path"),
     ("py:class", "subprocess.CompletedProcess"),
+    ("py:class", "subprocess.Popen"),
 ]