+++ tests/batcomputer.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+
+from batcomputer_cli.__main__ import CLI
+
+CLI()